requirements.txt
openai
chromadb
python-dotenv
docker
crewai































































































